# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Sora Store API - Digital products store
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  nameAr: string;
  slug: string;
  icon: string;
  color: string;
  productCount: number;
  brandIcons?: string[];
}

export interface Product {
  id: number;
  nameAr: string;
  /** @nullable */
  descriptionAr?: string | null;
  categoryId: number;
  categorySlug: string;
  price: number;
  originalPrice: number;
  discount: number;
  rating: number;
  reviewCount: number;
  image: string;
  currency: string;
  /** @nullable */
  isBestSeller?: boolean | null;
  /** @nullable */
  rank?: number | null;
  /** @nullable */
  period?: string | null;
  /** @nullable */
  isFlashSale?: boolean | null;
  /** @nullable */
  featured?: boolean | null;
  /** @nullable */
  badge?: string | null;
}

export interface FlashSaleResponse {
  products: Product[];
  endsAt: string;
}

export interface Review {
  id: number;
  authorName: string;
  rating: number;
  text: string;
  avatar: string;
  /** @nullable */
  emoji?: string | null;
}

export interface StoreStats {
  customers: string;
  orders: string;
  products: string;
  supportHours: string;
}

export interface BlogPost {
  id: number;
  titleAr: string;
  slug: string;
  excerptAr: string;
  contentAr: string;
  coverImage: string;
  authorName: string;
  /** @nullable */
  authorAvatar?: string | null;
  publishedAt: string;
  readingMins: number;
  tag: string;
  /** @nullable */
  views?: number | null;
}

export type ListProductsParams = {
/**
 * @nullable
 */
categoryId?: number | null;
/**
 * @nullable
 */
categorySlug?: string | null;
/**
 * @nullable
 */
featured?: boolean | null;
/**
 * @nullable
 */
limit?: number | null;
/**
 * price_asc | price_desc | rating | newest
 * @nullable
 */
sort?: string | null;
};

export type SearchProductsParams = {
q: string;
/**
 * @nullable
 */
limit?: number | null;
};

export type ListBlogPostsParams = {
/**
 * @nullable
 */
limit?: number | null;
/**
 * @nullable
 */
tag?: string | null;
};