# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Sora Store API - Digital products store
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: products
  - name: categories
  - name: flash-sales
  - name: reviews
  - name: stats
  - name: blog
  - name: search

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all product categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional filters
      parameters:
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: categorySlug
          in: query
          schema:
            type: ["string", "null"]
        - name: featured
          in: query
          schema:
            type: ["boolean", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: sort
          in: query
          schema:
            type: ["string", "null"]
            description: "price_asc | price_desc | rating | newest"
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/search:
    get:
      operationId: searchProducts
      tags: [search]
      summary: Search products by name
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/flash-sales:
    get:
      operationId: listFlashSaleProducts
      tags: [flash-sales]
      summary: Get current flash sale products
      responses:
        "200":
          description: Flash sale products with countdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlashSaleResponse"

  /products/best-sellers:
    get:
      operationId: listBestSellers
      tags: [products]
      summary: Get best-selling products
      responses:
        "200":
          description: Best seller products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found

  /products/{id}/related:
    get:
      operationId: getRelatedProducts
      tags: [products]
      summary: Get related products for a given product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Related products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List customer reviews
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

  /stats:
    get:
      operationId: getStoreStats
      tags: [stats]
      summary: Get store statistics
      responses:
        "200":
          description: Store stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreStats"

  /blog/posts:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List blog posts
      parameters:
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: tag
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of blog posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogPost"

  /blog/posts/{slug}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get a single blog post by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "404":
          description: Not found

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    Category:
      type: object
      required: [id, nameAr, slug, icon, color, productCount]
      properties:
        id:
          type: integer
        nameAr:
          type: string
        slug:
          type: string
        icon:
          type: string
        color:
          type: string
        productCount:
          type: integer
        brandIcons:
          type: array
          items:
            type: string

    Product:
      type: object
      required: [id, nameAr, categoryId, categorySlug, price, originalPrice, discount, rating, reviewCount, image, currency]
      properties:
        id:
          type: integer
        nameAr:
          type: string
        descriptionAr:
          type: ["string", "null"]
        categoryId:
          type: integer
        categorySlug:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        discount:
          type: integer
        rating:
          type: number
        reviewCount:
          type: integer
        image:
          type: string
        currency:
          type: string
        isBestSeller:
          type: ["boolean", "null"]
        rank:
          type: ["integer", "null"]
        period:
          type: ["string", "null"]
        isFlashSale:
          type: ["boolean", "null"]
        featured:
          type: ["boolean", "null"]
        badge:
          type: ["string", "null"]

    FlashSaleResponse:
      type: object
      required: [products, endsAt]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        endsAt:
          type: string
          format: date-time

    Review:
      type: object
      required: [id, authorName, rating, text, avatar]
      properties:
        id:
          type: integer
        authorName:
          type: string
        rating:
          type: integer
        text:
          type: string
        avatar:
          type: string
        emoji:
          type: ["string", "null"]

    StoreStats:
      type: object
      required: [customers, orders, products, supportHours]
      properties:
        customers:
          type: string
        orders:
          type: string
        products:
          type: string
        supportHours:
          type: string

    BlogPost:
      type: object
      required: [id, titleAr, slug, excerptAr, contentAr, coverImage, authorName, publishedAt, readingMins, tag]
      properties:
        id:
          type: integer
        titleAr:
          type: string
        slug:
          type: string
        excerptAr:
          type: string
        contentAr:
          type: string
        coverImage:
          type: string
        authorName:
          type: string
        authorAvatar:
          type: ["string", "null"]
        publishedAt:
          type: string
          format: date-time
        readingMins:
          type: integer
        tag:
          type: string
        views:
          type: ["integer", "null"]