# lib/api-zod/src/generated/types/blogPost.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Sora Store API - Digital products store
 * OpenAPI spec version: 0.2.0
 */

export interface BlogPost {
  id: number;
  titleAr: string;
  slug: string;
  excerptAr: string;
  contentAr: string;
  coverImage: string;
  authorName: string;
  /** @nullable */
  authorAvatar?: string | null;
  publishedAt: Date;
  readingMins: number;
  tag: string;
  /** @nullable */
  views?: number | null;
}