# lib/api-zod/src/generated/types/category.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Sora Store API - Digital products store
 * OpenAPI spec version: 0.2.0
 */

export interface Category {
  id: number;
  nameAr: string;
  slug: string;
  icon: string;
  color: string;
  productCount: number;
  brandIcons?: string[];
}