# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Sora Store API - Digital products store
 * OpenAPI spec version: 0.2.0
 */

export interface Product {
  id: number;
  nameAr: string;
  /** @nullable */
  descriptionAr?: string | null;
  categoryId: number;
  categorySlug: string;
  price: number;
  originalPrice: number;
  discount: number;
  rating: number;
  reviewCount: number;
  image: string;
  currency: string;
  /** @nullable */
  isBestSeller?: boolean | null;
  /** @nullable */
  rank?: number | null;
  /** @nullable */
  period?: string | null;
  /** @nullable */
  isFlashSale?: boolean | null;
  /** @nullable */
  featured?: boolean | null;
  /** @nullable */
  badge?: string | null;
}