# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Sora Store API - Digital products store
 * OpenAPI spec version: 0.2.0
 */

export interface Review {
  id: number;
  authorName: string;
  rating: number;
  text: string;
  avatar: string;
  /** @nullable */
  emoji?: string | null;
}