# lib/api-zod/src/generated/types/storeStats.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Sora Store API - Digital products store
 * OpenAPI spec version: 0.2.0
 */

export interface StoreStats {
  customers: string;
  orders: string;
  products: string;
  supportHours: string;
}